+++ aws/cloudformation/util.go
@@ -539,8 +539,9 @@ func WaitForStackOperationComplete(stackID string,
 	startTime := time.Now()
 
 	// Startup a spinner...
-	charSetIndex := 6
-	cliSpinner := spinner.New(spinner.CharSets[charSetIndex], 333*time.Millisecond)
+	charSetIndex := 7
+	cliSpinner := spinner.New(spinner.CharSets[charSetIndex],
+		333*time.Millisecond)
 	cliSpinnerStarted := false
 
 	// Poll for the current stackID state, and