+++ CHANGES.md
@@ -1,5 +1,37 @@
 # Change Notes
 
+## v1.9.0 - The LayerCake Edition 🍰
+
+- :warning: **BREAKING**
+- :checkered_flag: **CHANGES**
+  - Added `LambdaAWSInfo.Layers` field to support [Lambda Layers](https://docs.aws.amazon.com/lambda/latest/dg/configuration-layers.html)
+    - Usage:
+      ```go
+      lambdaRole := sparta.IAMRoleDefinition{
+        Privileges: []sparta.IAMRolePrivilege{
+          iamBuilder.Allow("lambda:GetLayerVersion").
+            ForResource().
+            Literal("*").
+            ToPrivilege(),
+        },
+      }
+      lambdaFn, lambdaFnErr := sparta.NewAWSLambda("Hello World",
+        helloWorld,
+        lambdaRole)
+      lambdaFn.Layers = []gocf.Stringable{
+        gocf.String("arn:aws:lambda:us-west-2:123412341234:layer:ffmpeg:1"),
+      }
+      ```
+  - Added `WithCondition` to [IAM Builder](https://godoc.org/github.com/mweagle/Sparta/aws/iam/builder)
+  - Added `s3Site.UserManifestData` map property to allow for custom user data to be included in _MANIFEST.json_ content that is deployed to an S3 Site bucket.
+    - Userdata is scoped to a **userdata** keyname in _MANIFEST.json_
+    - See the [SpartaAmplify](https://github.com/mweagle/SpartaAmplify) sample app for a complete example.
+  - Added `github.com/mweagle/Sparta/system.RunAndCaptureOSCommand`
+    - This is convenience function to support alternative `io.Writer` sinks for _stdout_ and _stderr_.
+  - Minor usability improvements to `--status` report output
+- :bug:  **FIXED**
+  - [overview page is broken](https://github.com/mweagle/Sparta/issues/133)
+
 ## v1.8.0 - The #postReInvent Edition ⌛️
 
 - :warning: **BREAKING**